--- splitmerge/worker/main.go
@@ -1,31 +1,0 @@
-package main
-
-import (
-	"log"
-
-	"go.temporal.io/sdk/client"
-	"go.temporal.io/sdk/worker"
-
-	"github.com/temporalio/samples-go/splitmerge"
-)
-
-func main() {
-	// The client and worker are heavyweight objects that should be created once per process.
-	c, err := client.Dial(client.Options{
-		HostPort: client.DefaultHostPort,
-	})
-	if err != nil {
-		log.Fatalln("Unable to create client", err)
-	}
-	defer c.Close()
-
-	w := worker.New(c, "split-merge", worker.Options{})
-
-	w.RegisterWorkflow(splitmerge.SampleSplitMergeWorkflow)
-	w.RegisterActivity(splitmerge.ChunkProcessingActivity)
-
-	err = w.Run(worker.InterruptCh())
-	if err != nil {
-		log.Fatalln("Unable to start worker", err)
-	}
-}
